requirements.txt
amqp==5.0.5
aniso8601==9.0.0
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
bandit==1.7.0
billiard==3.6.3.0
black==20.8b1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
flake8==3.8.4
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
gitdb==4.0.5
GitPython==3.1.14
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pbr==5.5.1
pluggy==0.13.1
prompt-toolkit==3.0.16
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
smmap==3.0.5
SQLAlchemy==1.3.23
stevedore==3.3.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
